refactor: split monolithic web-tools/index.ts into focused modules
- config.ts: provider registry, config load/save, API key resolution
- tool-impl.ts: standalone searchImpl/fetchImpl for testing
- image.ts: standalone imageImpl with download/resize logic
- tools.ts: tool registration, schemas, render/execute wrappers
- command.ts: /web-tools interactive settings command
- render.ts: shared collapsed text result renderer

Also strip Exa Summary: prefix in search descriptions, and stabilize
search-fetch test query for deterministic first-result fetch.
