+++ test/cases.json
@@ -8,7 +8,7 @@
 	{
 		"id": "search-fetch",
 		"tool": "search",
-		"args": { "query": "quantum computing explained", "limit": 2, "fetchResult": true },
+		"args": { "query": "NIST quantum computing explained", "limit": 2, "fetchResult": true },
 		"providers": ["firecrawl", "exa", "tavily"]
 	},
 	{
